Fix inference
